requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.28.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
pywavelets>=1.8.0
scikit-image>=0.25.0
matplotlib>=3.7.0
tensorflow>=2.12.0
tf2onnx>=1.16.0
onnxruntime>=1.15.0
